refactor: extract ResponsiveEditorView base for the mode-rail views
All five editor views (Atlas / Cast / Tiles / Scene-editor / Data) repeated the same responsive-shell machinery: the show-library / library-collapsed (+ show-inspector / inspector-collapsed) GObject properties with their camelCase get/set + backing fields, syncActiveMode, and the mode-changed signal — ~180 lines copy-pasted.

A new ResponsiveEditorView base (extends Adw.Bin) owns that contract; the views extend it, keep only their extras (show-quickview for Cast/Tiles, scene-name/project-name), and override the _onInspectorCollapsedChanged hook for per-view side-effects (Cast/Tiles hide the quick-view; Tiles also reparents its inspector). Each view's .blp template parent becomes $PixelRpgResponsiveEditorView (GTK requires the template parent to match the instance's actual superclass).

Behaviour-preserving; validated live via MCP (Cast desktop quick-view + phone rail-overlay/no-quickview, Tiles desktop inspector sidebar, Data view) with no GObject criticals.
