+++ apps/maker-gjs/src/widgets/atlas-view.blp
@@ -1,7 +1,7 @@
 using Gtk 4.0;
 using Adw 1;
 
-template $AtlasView : Adw.Bin {
+template $AtlasView : $PixelRpgResponsiveEditorView {
   // Two-level OverlaySplitView mirrors `scene-editor-view.blp` —
   // ModeRail (left) + SceneInspector (right) both span the full
   // window height. The middle pane is an `Adw.ToolbarView` with the
