+++ apps/maker-gjs/src/widgets/cast-view.blp
@@ -1,7 +1,7 @@
 using Gtk 4.0;
 using Adw 1;
 
-template $CastView : Adw.Bin {
+template $CastView : $PixelRpgResponsiveEditorView {
   // Outer OverlaySplitView holds the ModeRail (left navigation) for the
   // whole view's lifetime. The central content is an `Adw.NavigationView`
   // master-detail drill-down: the GALLERY page lists every character as a
