+++ apps/maker-gjs/src/widgets/index.ts
@@ -5,6 +5,7 @@ export * from './atlas-view'
 export * from './cast-view'
 export * from './data-view'
 export * from './preferences-dialog'
+export * from './responsive-editor-view'
 export * from './scene-editor-view'
 export * from './share-dialog'
 export * from './tiles-view'
