+++ apps/maker-gjs/src/widgets/scene-editor-view.blp
@@ -1,7 +1,7 @@
 using Gtk 4.0;
 using Adw 1;
 
-template $SceneEditorView : Adw.Bin {
+template $SceneEditorView : $PixelRpgResponsiveEditorView {
   Adw.OverlaySplitView outer_split {
     sidebar-position: start;
     min-sidebar-width: 220;
