+++ apps/maker-gjs/src/widgets/tiles-view.blp
@@ -1,7 +1,7 @@
 using Gtk 4.0;
 using Adw 1;
 
-template $TilesView : Adw.Bin {
+template $TilesView : $PixelRpgResponsiveEditorView {
   // Same master-detail shape as cast-view: ModeRail (left) holds the
   // whole view; the central `Adw.NavigationView` drills from a tileset
   // card GALLERY (with a desktop quick-view sidebar) into a per-tileset
